requirements.txt
astral>=3.2
PyYAML>=6.0.1
